az-204/cosmodb/requirements.txt
azure-cosmos
python-dotenv